fix: change entity IDs from Long to Integer to match database schema
- Database uses serial (INTEGER) but entities used Long (BIGINT)
- Updated all entities, repositories, DTOs, services, and controllers
- Updated all test files to use Integer IDs
- Improved component test error handling with detailed messages
- Resolves schema validation error: wrong column type encountered

+++ src/main/java/com/xqfitness/readservice/dto/WorkoutDayDetailDTO.java
@@ -13,8 +13,8 @@ import java.util.stream.Collectors;
 @NoArgsConstructor
 @AllArgsConstructor
 public class WorkoutDayDetailDTO {
-    private Long id;
-    private Long routineId;
+    private Integer id;
+    private Integer routineId;
     private Integer dayNumber;
     private String dayName;
     private String notes;
